sigma-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyyaml==6.0.2
pydantic==2.10.4
